Introduction to the Microsoft .NET platform. Basics of the C#
Works:
Task 1
The user enters from the keyboard a number in the range from 1 to 100. If the number is a multiple of 3 (divisible by 3 without
remainder) you need to display the word Fizz. If the number is a multiple of 5
you need to display the word Buzz. If the number is a multiple of 3 and 5 is necessary
bring out Fizz Buzz. If the number is not a multiple, 3 and 5 are not necessary
output the number itself.
If the user entered a value not in the range of 1
up to 100, an error message should be displayed.
Task 2
The user enters two numbers from the keyboard. First
the number is the value, the second number is the percentage, which
must be calculated. For example, we entered from the keyboard
90 and 10. It is necessary to display 10 percent of 90 on the screen.
Result: 9.
Task 3
The user enters four digits from the keyboard.
It is necessary to create a number containing these digits. For example, if 1, 5, 7, 8 is entered from the keyboard, then it is necessary
to form the number 1578.
Task 4
The user enters a six-digit number. After that
the user enters digit numbers to exchange digits.
For example, if the user entered one and six — this
it means that the first and sixth places need to be exchanged
numbers
The number 723895 should turn into 523897.
If the user has not entered a six-digit number, an error message must be displayed.
Task 5
The user enters the date from the keyboard. The application should display the name of the season and the day of the week.
For example, if 22.12.2021 is entered, the application should
will display Winter Wednesday.
Task 6
The user enters temperature readings from the keyboard. Depending on the user's choice, the program converts the temperature from Fahrenheit to Celsius
or vice versa.
Task 7
The user enters two numbers from the keyboard. It is necessary
show all even numbers in the specified range. If
the boundaries of the range are incorrectly specified, it is necessary to normalize the boundaries. For example, a user
entered 20 and 11, normalization is required, after which
the start of the range will be 11 and the end will be 20.